+++ domain/src/test/kotlin/fr/sacane/jmanager/domain/fake/FakeFactory.kt
@@ -106,6 +106,7 @@ object FakeFactory {
     val deleteTransactionsByIdsService = DeleteTransactionsByIdsService(transactionRepository, sessionManager, bookletRepository, manager, inMemoryTrackerRepository)
     val excludeVirtualTransactionService = ExcludeVirtualTransactionService(sessionManager, bookletRepository, inMemoryTrackerRepository)
     val confirmPreviewTransactionService = ConfirmPreviewTransactionService(transactionRepository, sessionManager, bookletRepository, manager)
+    val confirmVirtualTransactionService = ConfirmVirtualTransactionService(transactionRepository, sessionManager, bookletRepository, inMemoryTrackerRepository, inMemoryTagRepository, manager)
     val loginService = LoginService(userRepository, sessionManager, DefaultHasher, tokenGenerator)
     val logoutService = LogoutService(sessionManager)
     val refreshSessionService = RefreshSessionService(sessionManager, userRepository, tokenGenerator)